web-app/requirements.txt
flask==3.0.3
gunicorn==20.0.4
paho-mqtt==1.6.1
requests==2.31.0
